fix: soundboard on PostgreSQL, and generated validators without a local path
The sound list sorted with SQLite's COLLATE NOCASE, which PostgreSQL rejects;
it now sorts by LOWER(name). duration_ms and size are BIGINT on both engines
to match the i64 the Any driver decodes. validators.js had picked up a
symlinked node_modules path in two bundle comments.

+++ crates/paracord-db/migrations/20260924000201_soundboard_sounds.sql
@@ -12,9 +12,9 @@ CREATE TABLE IF NOT EXISTS soundboard_sounds (
     -- 0..100 percent, applied client-side on top of the listener's own
     -- soundboard volume.
     volume      INTEGER NOT NULL DEFAULT 100,
-    duration_ms INTEGER NOT NULL,
+    duration_ms BIGINT NOT NULL,
     content_type TEXT NOT NULL,
-    size        INTEGER NOT NULL,
+    size        BIGINT NOT NULL,
     asset_key   TEXT NOT NULL,
     creator_id  BIGINT REFERENCES users(id),
     created_at  TEXT NOT NULL DEFAULT (datetime('now'))
